requirements.txt
beautifulsoup4>=4.12
PyYAML>=6.0
pypdf>=5.0
